--- TrackPropagation/SteppingHelixPropagator/plugins/module.cc
@@ -1,10 +1,0 @@
-#include "SteppingHelixPropagatorESProducer.h"
-#include "FWCore/Framework/interface/EventSetup.h"
-#include "FWCore/Framework/interface/ESHandle.h"
-#include "FWCore/Framework/interface/ModuleFactory.h"
-#include "FWCore/Framework/interface/ESProducer.h"
-
-
-#include "FWCore/Utilities/interface/typelookup.h"
-
-DEFINE_FWK_EVENTSETUP_MODULE(SteppingHelixPropagatorESProducer);
